+++ kmcp/main.go
@@ -26,6 +26,7 @@ import (
 	"runtime"
 
 	colorable "github.com/mattn/go-colorable"
+	"github.com/pkg/profile"
 	"github.com/shenwei356/go-logging"
 	"github.com/shenwei356/kmcp/kmcp/cmd"
 )
@@ -46,7 +47,7 @@ func init() {
 
 func main() {
 	// go tool pprof -http=:8080 cpu.pprof
-	// defer profile.Start(profile.CPUProfile, profile.ProfilePath(".")).Stop()
+	defer profile.Start(profile.CPUProfile, profile.ProfilePath(".")).Stop()
 
 	// go tool trace -http=:8080 trace.out
 	// defer profile.Start(profile.TraceProfile, profile.ProfilePath(".")).Stop()